Moved loggings_handlers.py to new submodule loggers
