+++ bot/statalib/__init__.py
@@ -10,7 +10,8 @@ from .handlers import *
 from .subscriptions import *
 from .network import *
 from .permissions import *
-from .loggings_handlers import *
+
+from .loggers.handlers import *
 
 from .render.progress import *
 from .render.tools import *
@@ -45,6 +46,5 @@ __all__ = [
     'mcfetch',
     'subscriptions'
     'network',
-    'permissions',
-    'logging_handlers'
+    'permissions'
 ]
