+++ bot/statalib/loggers/__init__.py
@@ -1,0 +1,1 @@
+from .handlers import *
